==> 02_Classification_Notebooks/0_preprocess_data.ipynb <==
# 0_preprocessing the Crypto datasets
-------------------
Group , October 19, 2022
1. Eduardo Garcia
2. Nari Kim
3. Thi Anh Ba Dang
4. Vishnu Prabhakar
5. VS Chaitanya Madduri
6. Yumeng Zhang


> <i>Description: merging the daily and minute crypto datasets to respective single files.</i>


This mode

### Pre requisites: 
1. And add the shortcut of the drive link : https://drive.google.com/drive/folders/1KRMbTR4GNaDGlpBkRi3FcD_4ocbSULZ4 to your personal drive.


### Input files:
No files 

### Output files:

crypto_data_hour_cleaned_v2.csv
crypto_data_weekly_cleaned_v1.csv
crypto_data_minute_cleaned_v2.csv



DriveLink :https://drive.google.com/drive/folders/1F8P3UlqSE6lFpHyBidVArdsXJLenKi3O?usp=share_link




## 1. Import Required Packages 

import os
import pandas as pd
from datetime import datetime, timedelta

# Connecting to the google drive
from google.colab import drive
drive.mount('/content/drive')
from IPython.display import clear_output

def fill_empty_dates(dataframe_crypto):
  '''
  The funtion will add the missing dates to the crypto coins.
  input: 
  dataframe_crypto : The dataset where the missing dates to be added.
  Output: 
  final_df: date filled dataset.
  
  '''
  start_date = dataframe_crypto['Open Time'].min() 
  end_date = dataframe_crypto['Open Time'].max() 
  final_list = pd.date_range(start_date,end_date, 
                freq='min').strftime("%Y-%m-%d %H:%M:%S").tolist()
  date_df = pd.DataFrame(final_list, columns=['Open Time'])
  date_df['Open Time'] = pd.to_datetime(date_df['Open Time'])
  final_df = dataframe_crypto.merge(date_df, how='right', on='Open Time')
  final_df.ffill(inplace=True)
  return final_df

# last one yer records are made as test and rest of the records Train
def assign_train_test_tag(df):
  '''
  Assigning the test tag to the last 365 days entries and rest as Train
  to the crypto datasets
  input: 
  df : The dataset where the train/test tag need to be added.
  Output: 
  df: Data filled dataset.
  '''

  df['train_test'] = 'Train'
  df['train_test'][(df['Open Time']>max(df['Open Time']) -  timedelta(days=365) ) & (df['Open Time']<=max(df['Open Time']))] ='Test'
  df['train_test'][~((df['Open Time']>max(df['Open Time']) -  timedelta(days=365) ) & (df['Open Time']<=max(df['Open Time'])))] ='Train'
  
  
  return df 

def extract_hourly_data(crypto_data):
  '''
  The function processes the extarct the time series values from a given hourly range.
  Input:
  crypto_data: The minute data of the time series coin/stock.
  Output:
  temp_df : the final dataset will all the columns.
  '''
  crypto_data.sort_values(['Open Time'], inplace=True)
  crypto_data['Open Time'] = pd.to_datetime(crypto_data['Open Time'])
  crypto_data['temp_col'] = crypto_data['Open Time'].dt.date.astype(str) + '_' + crypto_data['Open Time'].dt.hour.astype(str)
  temp_df = crypto_data.groupby('temp_col').agg({
      'Open Time':'min',
      'Open':'first',
      'High':'max',
      'Low':'min', 
      'Close': 'last', 
      'Volume': 'sum',
      
                                      })
  temp_df = temp_df.reset_index(drop=True).sort_values('Open Time')
  temp_df['Open Time'] = temp_df['Open Time'].dt.date.astype(str) + ' ' + temp_df['Open Time'].dt.hour.astype(str) +':00:00'
  temp_df['Open Time'] = pd.to_datetime(temp_df['Open Time'])
  return temp_df

def extract_daily_data(crypto_data):
   '''
  The function processes the extarct the time series values from a given daily range.
  Input:
  crypto_data: The minute/hourly data of the time series coin/stock.
  Output:
  temp_df : the final dataset will all the columns.
  '''
  crypto_data.sort_values(['Open Time'], inplace=True)
  crypto_data['Open Time'] = pd.to_datetime(crypto_data['Open Time'])
  crypto_data['temp_col'] = crypto_data['Open Time'].dt.date.astype(str) + '_' + crypto_data['Open Time'].dt.day.astype(str)
  temp_df = crypto_data.groupby('temp_col').agg({
      'Open Time':'min',
      'Open':'first',
      'High':'max',
      'Low':'min', 
      'Close': 'last', 
      'Volume': 'sum',
                                      })
  temp_df = temp_df.reset_index(drop=True).sort_values('Open Time')
  temp_df['Open Time'] = temp_df['Open Time'].dt.date.astype(str) + ' 00:00:00'
  temp_df['Open Time'] = pd.to_datetime(temp_df['Open Time'])
  return temp_df

def extract_weekly_data(crypto_data):
   '''
  The function processes the extarct the time series values from a given weekly range.
  Input:
  crypto_data: The minute/daily data of the time series coin/stock.
  Output:
  temp_df : the final dataset will all the columns.
  '''
  crypto_data.sort_values(['Open Time'], inplace=True)
  crypto_data['Open Time'] = pd.to_datetime(crypto_data['Open Time'])
  crypto_data['year'] = crypto_data['Open Time'].dt.year
  crypto_data['week_number'] = crypto_data['Open Time'].dt.isocalendar().week
  crypto_data['temp_col'] = crypto_data['year'].astype(str) + '_' + crypto_data['week_number'].astype(str)
  temp_df = crypto_data.groupby('temp_col').agg({      
      'Open Time':'min',
      'year':'min',
      'week_number':'min',
      'Open':'min',
      'High':'max',
      'Low':'min', 
      'Close': 'max', 
      'Volume': 'sum',
                                      })
  temp_df.reset_index(drop=True, inplace=True)
  return temp_df

project_folder = "/content/drive/MyDrive/DL_2022_Assignment"

 # extracting the folder structure so that the dic will have all the csv locations
folder_dic = {}
files = []
for main_path in os.listdir(project_folder):
  
  sub_folder = project_folder + '/' +  main_path
  folder_dic[sub_folder] = []
  for path in os.listdir(sub_folder):
      # check if current path is a file
      if os.path.isfile(os.path.join(sub_folder, path)):
          folder_dic[sub_folder].append(path)

folder_dic.keys()

print("Number of the daily data files : {}".format(len(folder_dic['/content/drive/MyDrive/DL_2022_Assignment/Crypto_data_daily'])))


print("Number of the minute data files : {}".format(len(folder_dic['/content/drive/MyDrive/DL_2022_Assignment/Crypto_data_minute'])))

# making one dataframe out of the minute and the daily data 

folder_dic['/content/drive/MyDrive/DL_2022_Assignment/Crypto_data_minute'][2]

## Processing the Minute date extarcting Hourly, weekly , Minute data.

col_names=['Open Time',  'Open', 'High', 'Low', 'Close', 'Volume']

master_hour_df = pd.DataFrame(columns=col_names)
master_daily_df = pd.DataFrame(columns=col_names)
master_week_df = pd.DataFrame(columns=col_names)


minute_path = "/content/drive/MyDrive/DL_2022_Assignment/Crypto_data_minute"
for file_name  in folder_dic[minute_path]:
  df = pd.read_csv(minute_path + '/' +  file_name, names=col_names)
  df['Open Time'] = pd.to_datetime(df['Open Time'])
  # df = fill_empty_dates(df.copy())
   
  

  daily_df = extract_daily_data(df.copy())
  hour_df = extract_hourly_data(df.copy())
  weekly_df = extract_weekly_data(df.copy())


  
  daily_df = assign_train_test_tag(daily_df.copy())
  hour_df = assign_train_test_tag(hour_df.copy())
  weekly_df = assign_train_test_tag(weekly_df.copy())


  daily_df['Crypto'] = file_name.split("_")[0]
  hour_df['Crypto'] = file_name.split("_")[0]
  weekly_df['Crypto'] = file_name.split("_")[0]

  master_hour_df = pd.concat([master_hour_df, hour_df])
  master_daily_df = pd.concat([master_daily_df, daily_df])
  master_week_df = pd.concat([master_week_df, weekly_df])

## New columns

### Calculating the percentage change hourly data

# Calculating the percentage change

master_hour_df['pct_change_1hour'] = master_hour_df['Close'].pct_change(periods=1)
master_hour_df['pct_change_2hour'] = master_hour_df['Close'].pct_change(periods=2)
master_hour_df['pct_change_1day'] = master_hour_df['Close'].pct_change(periods=24)

master_hour_df.head(2)

### Hourly Data Sample

master_hour_df[['train_test','Crypto']].value_counts()

# exporting the dataframe to csv
folder_path = '/content/drive/MyDrive/MADS_23_DL_final_project'
master_hour_df.to_csv(folder_path + "/data/crypto_data_hour_cleaned_v2.csv", index=None)

### Daily Data Sample

master_daily_df[['train_test','Crypto']].value_counts()

# exporting the dataframe to csv
folder_path = '/content/drive/MyDrive/MADS_23_DL_final_project'
master_daily_df.to_csv(folder_path + "/data/crypto_data_daily_cleaned_v1.csv", index=None)

### Weekly data sample

master_week_df[['train_test','Crypto']].value_counts()

# exporting the dataframe to csv
folder_path = '/content/drive/MyDrive/MADS_23_DL_final_project'
master_week_df.to_csv(folder_path + "/data/crypto_data_weekly_cleaned_v1.csv", index=None)

### combining all the datasets into one csv

# exporting the dataframe to csv
folder_path = '/content/drive/MyDrive/MADS_23_DL_final_project'
master_minute_df.to_csv(folder_path + "/data/crypto_data_minute_cleaned_v2.csv", index=None)

## Merging the  Daily data

Discarded the code as we are not using the daily data for the analysis

col_names = ['Open Time', 'Open Time.1', 'Open', 'High', 'Low', 'Close', 'Volume',
       'Close Time', 'Quote Asset Volume', 'Number of Trades',
       'TB Base Volume', 'TB Quote Volume', 'Ignore']

master_daily_df = pd.DataFrame(columns=col_names)
minute_path = "/content/drive/MyDrive/DL_2022_Assignment/Crypto_data_daily"
for file_name  in folder_dic[minute_path]:
  df = pd.read_csv(minute_path + '/' +  file_name, names=col_names)
  df['filename_name'] = file_name
  master_daily_df = pd.concat([master_daily_df, df])

# exporting the dataframe to csv
folder_path = '/content/drive/MyDrive/MADS_23_DL_final_project'
master_daily_df.to_csv(folder_path + "/crypto_data_daily_all.csv", index=None)

master_daily_df.shape

## Loading the minute data

folder_path = '/content/drive/MyDrive/MADS_23_DL_final_project'
master_minute_df = pd.read_csv(folder_path + "/data/crypto_data_minute_cleaned_v2.csv")

master_minute_df.shape

master_minute_df['Open Time'] = pd.to_datetime(master_minute_df['Open Time'])
master_minute_df['Open_time_minute'] = master_minute_df['Open Time'].dt.minute
final_df = master_minute_df[master_minute_df['Open_time_minute'] == 59]

final_df.drop(['Open_time_minute'], axis=1).to_csv(folder_path + "/data/crypto_data_hour_cleaned_v1.csv", index=None)

final_df.shape

master_minute_df.sample(10)

## Crypto Split

folder_path = '/content/drive/MyDrive/MADS_23_DL_final_project'
master_minute_df = pd.read_csv(folder_path + "/data/crypto_data_daily_cleaned_v1.csv")

import plotly.express as px

df2 = master_minute_df.groupby(['Crypto'])['Crypto'].count().reset_index(name='count')


fig = px.pie(df2 , values='count',
             names='Crypto', 
             title='Crypto', 
             width=600, height=500,
             color_discrete_sequence=px.colors.sequential.RdBu)
fig.update_layout( title_x=0.5)
fig.show()

import plotly.express as px


fig = px.bar(df2.sort_values(by='count') , 
             x='Crypto', 
             y='count',
             title='Crypto', 
             width=600, height=500,
            #  color_discrete_sequence=px.colors.sequential.RdBu
             )
fig.update_layout( title_x=0.5)
fig.show()

folder_path = '/content/drive/MyDrive/MADS_23_DL_final_project/data/old_files'
master_minute_df = pd.read_csv(folder_path + "/crypto_data_minute_cleaned_v2.csv")

master_minute_df.shape

# file path
folder_path = '/content/drive/MyDrive/MADS_23_DL_final_project'
hour = pd.read_csv(folder_path + '/data/crypto_data_daily_cleaned_v1.csv')

hour.shape

## End of the Notebook
