==> 02_Classification_Notebooks/4_model_building_LSTM_F13.ipynb <==
4_model_building_LSTM_F13
-------------------
Group , November 18, 2022
1. Eduardo Garcia
2. Nari Kim
3. Thi Anh Ba Dang
4. Vishnu Prabhakar
5. VS Chaitanya Madduri
6. Yumeng Zhang


> <i>Description: Applying the LSTM model on the daily data.</i>


- We are builing a basic model using the columns and the target provided in the intial dataset.

- Used Feature Set 13 which are basic features and we have calculated the Lag and rolled data.


### Pre requisites: 
1. And add the shortcut of the drive link : https://drive.google.com/drive/folders/1F8P3UlqSE6lFpHyBidVArdsXJLenKi3O?usp=share_link to your personal drive.


Files:
crypto_data_hour_cleaned_v2.csv - Hourly Data

### Output files:

Files:LSTM_tw28.h5



# install packages
! pip install keras-tuner --upgrade
! pip install ta
! pip install seglearn

# Load and transform data


# Connecting to the google drive
from google.colab import drive
drive.mount('/content/drive')
from IPython.display import clear_output



import pandas as pd
import warnings
warnings.filterwarnings("ignore")
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np

from sklearn.model_selection import RandomizedSearchCV

#picking models for prediction.
from sklearn.svm import SVC

# file path
folder_path = '/content/drive/MyDrive/MADS_23_DL_final_project'
daily = pd.read_csv(folder_path + '/data/crypto_data_daily_cleaned_v1.csv')

daily.head()

daily

## Train / Test Split

daily['train_test'].value_counts(normalize=True)

df = daily.copy()

# train test split
train_df = df[df['train_test']=='Train']
test_df = df[df['train_test']=='Test']

train_df.head()

train_df.columns

## Calculate percentage change for Train and Test

def calculate_pct_change(df):
  coins = df.Crypto.unique()
  df_pct_change = pd.DataFrame()
  for coin in coins:
    x = df[df['Crypto']==coin]
    x['pct_change_1day'] = x['Close'].pct_change(1)
    df_pct_change = pd.concat([df_pct_change,x])
  return df_pct_change

train_df = calculate_pct_change(train_df)
test_df = calculate_pct_change(test_df)

test_df.head()

## Generate lag features

# function to calculate the market capitalization of coins at each time point
def create_market_volumn_features(df):
  
  # calculate value of each cryto at certain time points 
  df['Total_Value'] = df['Close']*df['Volume']
  # the sum of values at each time point
  sum_at_timepoints = df.groupby('Open Time').sum()['Total_Value']
  merged = df.merge(sum_at_timepoints, how='left',
                              on='Open Time', suffixes=('','_market'))
  merged['Value_Weight'] = merged['Total_Value']/merged['Total_Value_market']
    
  return merged

# function to create 1 day lag
def create_shift_features(df, col = 'pct_change_1day'):
   df['1_d_lag'] = df[col].shift(periods=1)
  #  df['1_w_lag'] = df[col].shift(periods=7)
   return df

# stack all functions together and iterate through all coins
from ta import add_all_ta_features
def create_analysis_colums(df):

  master_df = pd.DataFrame()
  crypto_coins = df['Crypto'].unique()

  for coin in crypto_coins:
    
    temp_df = df[df['Crypto']==coin]
    temp_df['pct_change_1day'] = temp_df['Close'].pct_change()
    temp_df = create_shift_features(temp_df.copy(),col = 'pct_change_1day')
    temp_df = add_all_ta_features(temp_df.copy(), open="Open", high="High", low="Low", close="Close", volume="Volume", fillna=True)
    if master_df.empty :
      master_df = temp_df
    else:
      master_df = pd.concat([master_df, temp_df])
  return master_df

# apply all functions
train_df =create_analysis_colums(train_df)
test_df =create_analysis_colums(test_df)

# calculate the market capitalization
train_df =create_market_volumn_features(train_df.copy())
test_df =create_market_volumn_features(test_df.copy())

# features selected from previous model importance analysis
feat_to_keep = ['Open Time', 'Open', 'High', 'Low', 'Close', 'Volume', 'train_test',
       'Crypto', 'pct_change_1day','volatility_kcw','trend_cci','volume_adi','momentum_ppo_hist','momentum_stoch','volatility_kcp','volume_em',
             'volatility_dcw','volume_vpt','volatility_bbw','Total_Value', 'Total_Value_market', 'Value_Weight']

lag_cols = [col for col in train_df.columns if 'lag' in col]
feat_to_keep.extend(lag_cols)

# only keep above features
train_df = train_df[feat_to_keep]
test_df = test_df[feat_to_keep]

# function to shift all technical features by 1 because it causes leakage
def shift_vol(df):
  impo_feat = ['volatility_kcw','trend_cci','volume_adi','momentum_ppo_hist','momentum_stoch','volatility_kcp','volume_em',
              'volatility_dcw','volume_vpt','volatility_bbw']
  master_df = pd.DataFrame()
  crypto_coins = df['Crypto'].unique()


  for coin in crypto_coins:
    temp_df = df[df['Crypto']==coin]
    for feat in impo_feat:    

      temp_df[feat] = temp_df[feat].shift(1)
    if master_df.empty :
      master_df = temp_df
    else:
      master_df = pd.concat([master_df, temp_df])
  return master_df

# apply shift function
train_df = shift_vol(train_df.copy())
test_df = shift_vol(test_df.copy())

## Extract year, month, day, hour and weekday from time stamp

### Encoding of ordinals


def encode_cyclicals(df_x):

    df_x['month_sin'] = np.sin(2*np.pi*df_x.month/12)
    df_x['month_cos'] = np.cos(2*np.pi*df_x.month/12)
    df_x.drop('month', axis=1, inplace=True)
    
    df_x['day_sin'] = np.sin(2*np.pi*df_x.day/31)
    df_x['day_cos'] = np.cos(2*np.pi*df_x.day/31)
    df_x.drop('day', axis=1, inplace=True)

    df_x['dayofweek_sin'] = np.sin(2*np.pi*df_x.weekday/7)
    df_x['dayofweek_cos'] = np.cos(2*np.pi*df_x.weekday/7)
    df_x.drop('weekday', axis=1, inplace=True)
    
    df_x['hour_sin'] = np.sin(2*np.pi*df_x.hour/24)
    df_x['hour_cos'] = np.cos(2*np.pi*df_x.hour/24)
    df_x.drop('hour', axis=1, inplace=True)

    df_x['hour_sin'] = np.sin(2*np.pi*df_x.minute/60)
    df_x['hour_cos'] = np.cos(2*np.pi*df_x.minute/60)
    df_x.drop('minute', axis=1, inplace=True)
    

    return df_x

# function to extract time features and encode them
def date_values_extraction(new_df):
  df = new_df.copy()
  df['year'] = pd.DatetimeIndex(df['Open Time']).year
  df['month'] = pd.DatetimeIndex(df['Open Time']).month
  df['day'] = pd.DatetimeIndex(df['Open Time']).day
  df['weekday'] = pd.DatetimeIndex(df['Open Time']).dayofweek


  df['Open Time'] = pd.to_datetime(df['Open Time'])
  df['minute'] = df['Open Time'].dt.minute
  df['hour'] = df['Open Time'].dt.hour
  df = encode_cyclicals(df.copy())
  return df

# apply above function
train_df = date_values_extraction(train_df)
test_df = date_values_extraction(test_df)

## One hot coding the coins




# Applying one hot encoding on Crypto Coin 
def crypto_one_hot_encoding(df):
  y_dummies = pd.get_dummies(df['Crypto'], prefix='Crypto', drop_first= False)
  # creating a additional column if the model is used for new coin.
  y_dummies['other_crypto'] =0
  return y_dummies

# roll data for one hot encoded dummy variables
def rolling_hot_data(train_df,TIME_WINDOW):
  train_hot = pd.DataFrame()
  for col in train_df.columns:
    if train_hot.empty:
        train_hot = train_df[train_df[col]==1][:-TIME_WINDOW]      
    else:
      train_hot = pd.concat([train_hot, train_df[train_df[col]==1][:-TIME_WINDOW]], axis=0)
  return train_hot
      

# one hot encoding the coins
train_onehot_data = crypto_one_hot_encoding(train_df)
test_onehot_data = crypto_one_hot_encoding(test_df)

# roll the one hot encoded data
TIME_WINDOW=28
train_onehot_data = rolling_hot_data(train_onehot_data,TIME_WINDOW)
test_onehot_data = rolling_hot_data(test_onehot_data,TIME_WINDOW)

test_onehot_data.shape

train_df['pct_change_1day'].describe()

test_df['pct_change_1day'].describe()

# Defining the Target Variable

We devide the data into 3 classes, one is with return below 0, one is above 0 but below market rate of return, one is above market rate of return. The market rate of return we use here is annual return of S&P 500 index in 2021.

# function to create classes
def create_target(df, prob='Classification'):
  if prob == 'Classification':
    market_RoR = 26.89
    market_RoR_1d = market_RoR/365
    df['Target'] = np.where(df['pct_change_1day']>0, 1,0)
    df['Target'] = np.where(df['pct_change_1day']>market_RoR_1d, 2,1)
    df['Target'][df['Target']==1] = np.where(df['pct_change_1day'][df['Target']==1]>=0, 1,0)
  elif prob == 'Regression': 
    df['Target'] = df['pct_change_1day']
  return df

# apply function
train_df = create_target(train_df, 'Classification')
test_df = create_target(test_df, 'Classification')

train_df['Target'].value_counts(normalize=True)

test_df['Target'].value_counts(normalize=True)

test_df.drop(['pct_change_1day'], axis=1, inplace=True) # droppping the column as we already extracted the target
train_df.drop(['pct_change_1day'], axis=1, inplace=True) # droppping the column as we already extracted the target

train_df.shape

test_df.shape

# define the targets
target = train_df['Target']
test_target = test_df['Target']

# Drop  columns

# drop unuseful columns
train_df.drop(['Target','Open Time','train_test',],axis=1,inplace=True)
test_df.drop(['Target','Open Time','train_test',],axis=1,inplace=True)

# dropping the list of the columns 

drop_columns = ['Open','Close']


if drop_columns:
  norm_train_df = train_df.drop(drop_columns,axis=1)
  norm_test_df = test_df.drop(drop_columns,axis=1)
else:
  norm_train_df = train_df
  norm_test_df = test_df

norm_train_df.columns

norm_train_df.shape

target.shape

norm_train_df.head()

norm_test_df.head()

# Roll the train and test data


# use past 28 days to forecast the next day
TIME_WINDOW=28
FORECAST_DISTANCE=1

from seglearn.transform import FeatureRep, SegmentXYForecast, last

# function to roll data
def Segment_multi(train_df, target):
  master_df_x = pd.DataFrame()
  master_df_y = pd.DataFrame()

  crypto_coins = df['Crypto'].unique()

  segmenter = SegmentXYForecast(width=TIME_WINDOW, step=1, y_func=last, forecast=FORECAST_DISTANCE)
  for coin in crypto_coins:

    coin_index = train_df[train_df['Crypto']==coin].index
    X_train_rolled, y_train_rolled,_=segmenter.fit_transform([train_df.iloc[coin_index].drop(['Crypto'], axis=1).values],[target.iloc[coin_index].values])
    
  
    if coin == 'BTC' :
      master_df_x = X_train_rolled
      master_df_y = y_train_rolled
    else:
      master_df_x = np.concatenate([master_df_x, X_train_rolled], axis=0)

      master_df_y = np.concatenate([master_df_y, y_train_rolled], axis=0)
    
  return master_df_x, master_df_y

# roll train and test data
train_roll, y_roll =  Segment_multi(norm_train_df.fillna(0), target)
test_roll, y_test_roll =  Segment_multi(norm_test_df.fillna(0), test_target)

norm_train_df.columns

train_roll.shape

test_roll.shape

# Evaluation function

from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
def generate_model_report(y_actual, y_predicted, metric_type):
    print("==================Printing the {} metrics===================".format(metric_type))
    if metric_type=='micro':
      print("Accuracy = " , round(accuracy_score(y_actual, y_predicted),4))
    print("Precision = " ,round(precision_score(y_actual, y_predicted, average=metric_type),4))
    print("Recall = " ,round(recall_score(y_actual, y_predicted, average=metric_type),4))
    print("F1 Score = " ,round(f1_score(y_actual, y_predicted, average=metric_type),4))
    print("=====================================")

# stateless LSTM model

from tensorflow import keras
from kerastuner.tuners import RandomSearch
from tensorflow.keras.layers import Dense, Dropout, LSTM, Bidirectional, BatchNormalization, Input, concatenate
from tensorflow.keras.models import Model
from tensorflow.keras import backend as be
from tensorflow.keras.callbacks import ModelCheckpoint, EarlyStopping
import keras_tuner

# use 2 inputs because LSTM cannot process non-sequential data
# so we use another dense layer to handle coin dummy data
float_input = Input(shape=(train_roll.shape[1],train_roll.shape[2]))
one_hot_input = Input(shape=(11,) )

first_lstm = LSTM(60)(float_input)#,stateful=True
# we also tried stateful/bidirectional LSTM but no improvement on the result
first_dense = Dense(50)(one_hot_input)
merge_one = concatenate([first_dense, first_lstm])
dense_inner = Dense(10)(merge_one)
dense_output = Dense(3, activation='softmax')(dense_inner)

model = Model(inputs=[float_input, one_hot_input], outputs=dense_output)
model.compile(loss='sparse_categorical_crossentropy', # we select it because it's multi-class classification
              optimizer='Adam',
              metrics=['sparse_categorical_crossentropy'])
#model.summary()
model.fit([train_roll,train_onehot_data], y_roll, epochs=20, 
          validation_data = ([test_roll,test_onehot_data], y_test_roll))#batch_size=BATCH_SIZE,

nn_pred=model.predict([test_roll,test_onehot_data])

# take argmax in each roll
df = pd.DataFrame(nn_pred)
nn_pred = df.idxmax(axis=1)

# model reports
generate_model_report(y_test_roll, nn_pred, 'micro')
generate_model_report(y_test_roll, nn_pred, 'macro')
generate_model_report(y_test_roll, nn_pred, 'weighted')

from tensorflow import keras
from kerastuner.tuners import BayesianOptimization, RandomSearch
from tensorflow.keras import Sequential
from tensorflow.keras.layers import Dense, Dropout, LSTM, Bidirectional, BatchNormalization
from tensorflow.keras import backend as be
from tensorflow.keras.callbacks import ModelCheckpoint, EarlyStopping
import keras_tuner


column_count=len(norm_train_df.columns)
TIME_WINDOW=28
DROPOUT_RATE=0.3
BATCH_SIZE = 32

# function to wrap the model
def build_model(hp):

    be.clear_session()

    float_input = Input(shape=(train_roll.shape[1],train_roll.shape[2] ))
    one_hot_input = Input(shape=(11,) )



    first_lstm = LSTM(units=hp.Int('units',min_value=32,
                                    max_value=64,
                                    step=2))(float_input)
    first_dense = Dense(50)(one_hot_input)
    merge_one = concatenate([first_lstm, first_dense])
    dense_inner = Dense(10)(merge_one)
    dense_output = Dense(3, activation='softmax')(dense_inner)



    model = Model(inputs=[float_input, one_hot_input], outputs=dense_output)
    model.compile(loss='sparse_categorical_crossentropy',
                  optimizer='Adam',
                  metrics=['sparse_categorical_accuracy'])

    return model

# run random search on the hyperparameters
random_opt_tuner = RandomSearch(
    build_model,
    objective=keras_tuner.Objective('val_sparse_categorical_accuracy','max'),
    # we tried to optimize on f1_macro, but nothing worked, the model didn't learn with customized f1 function
    max_trials=10,
    executions_per_trial=1,
    project_name='kerastuner_random_poc',
    overwrite=True)

random_opt_tuner.search([train_roll,train_onehot_data], y_roll, epochs=20, batch_size=BATCH_SIZE,
                    validation_data = ([test_roll,test_onehot_data], y_test_roll),verbose=1)#validation_split=0.2,


random_opt_model_best_model = random_opt_tuner.get_best_models(num_models=1)
model = random_opt_model_best_model[0]

from tensorflow.keras.callbacks import EarlyStopping

# fit the untuned model 
model_saver = ModelCheckpoint('/content/drive/MyDrive/MADS_23_DL_final_project/data/model_files/LSTM_daily', 
                            save_weights_only=True,
                            monitor='val_sparse_categorical_accuracy',
                            mode='max',
                            save_best_only=True)
early_stopping = EarlyStopping(monitor='val_sparse_categorical_accuracy', min_delta= 0.001, patience=5, baseline= 0.5)
model.fit([train_roll,train_onehot_data], 
          y_roll, 
          epochs=50, 
          validation_data = ([test_roll, test_onehot_data], y_test_roll),
          batch_size=BATCH_SIZE,
          callbacks=[early_stopping,model_saver])

# predict with test data and get the argmax
lstm_pred = model.predict([test_roll, test_onehot_data])
df = pd.DataFrame(lstm_pred)
lstm_pred = df.idxmax(axis=1)

# model reports
# there's a little bit increase of F1 macro
generate_model_report(y_test_roll, lstm_pred, 'micro')
generate_model_report(y_test_roll, lstm_pred, 'macro')
generate_model_report(y_test_roll, lstm_pred, 'weighted')

# this is actually the best LSTM we have but didn't get to save the notebook
from tensorflow.keras.models import load_model
model = load_model('/content/drive/MyDrive/MADS_23_DL_final_project/data/model_files/LSTM_daily/LSTM_tw28.h5')

# predict and get argmax
lstm_pred = model.predict([test_roll, test_onehot_data])
df = pd.DataFrame(lstm_pred)
lstm_pred = df.idxmax(axis=1)

# model reposts
generate_model_report(y_test_roll, lstm_pred, 'micro')
generate_model_report(y_test_roll, lstm_pred, 'macro')
generate_model_report(y_test_roll, lstm_pred, 'weighted')

## End of notebook
