==> 02_Classification_Notebooks/5_model_Testing_F13.ipynb <==
5_model_building_tuning_F13
-------------------
Group , November 18, 2022
1. Eduardo Garcia
2. Nari Kim
3. Thi Anh Ba Dang
4. Vishnu Prabhakar
5. VS Chaitanya Madduri
6. Yumeng Zhang


> <i>Description: We will compare different model outputs .</i>


- We are builing a basic model using the columns and the target provided in the intial dataset.

- Used Feature Set 13 which are basic features the data has.

### Pre requisites: 
1. And add the shortcut of the drive link : https://drive.google.com/drive/folders/1F8P3UlqSE6lFpHyBidVArdsXJLenKi3O?usp=share_link to your personal drive.


Files:
crypto_data_hour_cleaned_v2.csv - Hourly Data

### Output files:

Files:



! pip install ta
! pip install seglearn

# Load and transform data


# Connecting to the google drive
from google.colab import drive
drive.mount('/content/drive')
from IPython.display import clear_output



import pandas as pd
import warnings
warnings.filterwarnings("ignore")
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np

from sklearn.model_selection import RandomizedSearchCV

#picking models for prediction.
from sklearn.svm import SVC

# file path
folder_path = '/content/drive/MyDrive/MADS_23_DL_final_project'
daily = pd.read_csv(folder_path + '/data/crypto_data_daily_cleaned_v1.csv')

daily.head()

# Train / Test Split

daily['train_test'].value_counts(normalize=True)

df = daily.copy()

# train test split
train_df = df[df['train_test']=='Train']
test_df = df[df['train_test']=='Test']

train_df.head()

train_df.columns

## Code to normalization.

# from sklearn.preprocessing import MinMaxScaler
# import warnings
# from sklearn.exceptions import DataConversionWarning
# warnings.filterwarnings(action='ignore', category=DataConversionWarning)

# def minmax_scale(df_x, normalizers=None):


#     #Try to pass the column values as a parameter from outside 
#     features_to_minmax = ['year','Open','High', 'Low','Close','2h_lag', 'MovingAvg_2hour',
#        'ExpMovingAvg_2hour', '4h_lag', 'MovingAvg_4hour', 'ExpMovingAvg_4hour',
#        '6h_lag', 'MovingAvg_6hour', 'ExpMovingAvg_6hour', '8h_lag',
#        'MovingAvg_8hour', 'ExpMovingAvg_8hour', '10h_lag', 'MovingAvg_10hour',
#        'ExpMovingAvg_10hour', '12h_lag', 'MovingAvg_12hour',
#        'ExpMovingAvg_12hour', '14h_lag', 'MovingAvg_14hour',
#        'ExpMovingAvg_14hour', '16h_lag', 'MovingAvg_16hour',
#        'ExpMovingAvg_16hour', '18h_lag', 'MovingAvg_18hour',
#        'ExpMovingAvg_18hour', '20h_lag', 'MovingAvg_20hour',
#        'ExpMovingAvg_20hour', 'Total_Value',
#        'Total_Value_market']

#     if not normalizers:
#         normalizers = {}

#     for feat in features_to_minmax:
#         if feat not in normalizers:
#             normalizers[feat] = MinMaxScaler()
#             normalizers[feat].fit(df_x[feat].values.reshape(-1, 1))
        
#         df_x[feat] = normalizers[feat].transform(df_x[feat].values.reshape(-1, 1))

#     # series_y=normalizers["pct_change_2hour"].transform(series_y.values.reshape(-1, 1))

#     return df_x , normalizers

# norm_train_df, normalizers = minmax_scale(train_df)
# norm_test_df, _ = minmax_scale(test_df)

# exporting the dataframe to csv
# folder_path = '/content/drive/MyDrive/MADS_23_DL_final_project'
# master_minute_df.to_csv(folder_path + "/data/crypto_data_minute_cleaned_v2.csv", index=None)

# norm_train_df.head()

# Calculate percentage change

def calculate_pct_change(df):
  coins = df.Crypto.unique()
  df_pct_change = pd.DataFrame()
  for coin in coins:
    x = df[df['Crypto']==coin]
    x['pct_change_1day'] = x['Close'].pct_change(1)
    df_pct_change = pd.concat([df_pct_change,x])
  return df_pct_change

train_df = calculate_pct_change(train_df)
test_df = calculate_pct_change(test_df)

test_df.head()

# Generate new features

### Lag features, moving average, exponential moving average and market cap

def create_market_volumn_features(df):
  
  # calculate value of each cryto at certain time points 
  df['Total_Value'] = df['Close']*df['Volume']
  # the sum of values at each time point
  sum_at_timepoints = df.groupby('Open Time').sum()['Total_Value']
  merged = df.merge(sum_at_timepoints, how='left',
                              on='Open Time', suffixes=('','_market'))
  merged['Value_Weight'] = merged['Total_Value']/merged['Total_Value_market']
    
  return merged

def create_shift_features(df, col = 'pct_change_1day',lags=10, freq='daily'):

  if freq=='daily':
    mul_fact = 1
    symbol = 'd'
  elif freq=='weekly':
    mul_fact = 7
    symbol = 'W'
  elif freq=='monthly':
    mul_fact = 31
    symbol = 'mon'
  else:
    # setting default to daily 
    mul_fact = 1
    symbol = 'd'

  for iterator in range(1,lags+1):
    
    df['{}_{}_lag'.format(iterator, symbol)] = df[col].shift(periods=iterator*mul_fact)
    # df.loc[:,"Volatility_{}_{}".format(iterator, symbol)] = df[col].rolling(iterator*mul_fact).std().shift(1)
    

  return df

#list to collect all relevant lags 
from ta import add_all_ta_features
def create_analysis_colums(df):

  master_df = pd.DataFrame()
  crypto_coins = df['Crypto'].unique()

  for coin in crypto_coins:
    
    temp_df = df[df['Crypto']==coin]
    temp_df['pct_change_1day'] = temp_df['Close'].pct_change()
    
    # temp_df = create_shift_features(temp_df.copy(),col = 'pct_change_1day',lags=5, freq='weekly')
    temp_df = create_shift_features(temp_df.copy(),col = 'pct_change_1day',lags=2, freq='monthly')
    temp_df = create_shift_features(temp_df.copy(),col = 'pct_change_1day',lags=4, freq='weekly')
    temp_df = create_shift_features(temp_df.copy(),col = 'pct_change_1day',lags=12, freq='daily')
    temp_df = add_all_ta_features(temp_df.copy(), open="Open", high="High", low="Low", close="Close", volume="Volume", fillna=True)
    if master_df.empty :
      master_df = temp_df
    else:
      master_df = pd.concat([master_df, temp_df])
  return master_df

train_df =create_analysis_colums(train_df)
test_df =create_analysis_colums(test_df)

train_df.columns

train_df =create_market_volumn_features(train_df.copy())
test_df =create_market_volumn_features(test_df.copy())

train_df.columns[:50]

feat_to_keep = ['Open Time', 'Open', 'High', 'Low', 'Close', 'Volume', 'train_test',
       'Crypto', 'pct_change_1day','volatility_kcw','trend_cci','volume_adi','momentum_ppo_hist','momentum_stoch','volatility_kcp','volume_em',
             'volatility_dcw','volume_vpt','volatility_bbw','Total_Value', 'Total_Value_market', 'Value_Weight']
lag_cols = [col for col in train_df.columns if 'lag' in col]
#volatility_cols = [col for col in train_df.columns if 'Volatility' in col]
feat_to_keep.extend(lag_cols)
#feat_to_keep.extend(volatility_cols)

len(feat_to_keep)

train_df = train_df[feat_to_keep]
test_df = test_df[feat_to_keep]

len(train_df.columns)

def shift_vol(df):
  impo_feat = ['volatility_kcw','trend_cci','volume_adi','momentum_ppo_hist','momentum_stoch','volatility_kcp','volume_em',
              'volatility_dcw','volume_vpt','volatility_bbw']
  master_df = pd.DataFrame()
  crypto_coins = df['Crypto'].unique()


  for coin in crypto_coins:
    temp_df = df[df['Crypto']==coin]
    for feat in impo_feat:    

      temp_df[feat] = temp_df[feat].shift(1)
    if master_df.empty :
      master_df = temp_df
    else:
      master_df = pd.concat([master_df, temp_df])
  return master_df

train_df = shift_vol(train_df.copy())
test_df = shift_vol(test_df.copy())

# Extract year, month, day, hour and weekday from time stamp

### Encoding of ordinals


def encode_cyclicals(df_x):
  '''
  The function converts the date features encoded in the Sine and cosines.
  Input : 
  df_x : Input data frame to be processed
  Output :
  df_x : processed dataframe.
  '''
    #"month","day","hour", "minute", "dayofweek"
    
    

    df_x['month_sin'] = np.sin(2*np.pi*df_x.month/12)
    df_x['month_cos'] = np.cos(2*np.pi*df_x.month/12)
    df_x.drop('month', axis=1, inplace=True)
    
    df_x['day_sin'] = np.sin(2*np.pi*df_x.day/31)
    df_x['day_cos'] = np.cos(2*np.pi*df_x.day/31)
    df_x.drop('day', axis=1, inplace=True)

    df_x['dayofweek_sin'] = np.sin(2*np.pi*df_x.weekday/7)
    df_x['dayofweek_cos'] = np.cos(2*np.pi*df_x.weekday/7)
    df_x.drop('weekday', axis=1, inplace=True)
    
    df_x['hour_sin'] = np.sin(2*np.pi*df_x.hour/24)
    df_x['hour_cos'] = np.cos(2*np.pi*df_x.hour/24)
    df_x.drop('hour', axis=1, inplace=True)

    df_x['hour_sin'] = np.sin(2*np.pi*df_x.minute/60)
    df_x['hour_cos'] = np.cos(2*np.pi*df_x.minute/60)
    df_x.drop('minute', axis=1, inplace=True)
    
   
    
    return df_x

def date_values_extraction(new_df):
  df = new_df.copy()
  df['year'] = pd.DatetimeIndex(df['Open Time']).year
  df['month'] = pd.DatetimeIndex(df['Open Time']).month
  df['day'] = pd.DatetimeIndex(df['Open Time']).day
  df['weekday'] = pd.DatetimeIndex(df['Open Time']).dayofweek


  df['Open Time'] = pd.to_datetime(df['Open Time'])
  df['minute'] = df['Open Time'].dt.minute
  df['hour'] = df['Open Time'].dt.hour
  df = encode_cyclicals(df.copy())
  return df

train_df = date_values_extraction(train_df)
test_df = date_values_extraction(test_df)

## One hot coding the coins

By this process we are tagging which record belongs which coin

# Applying one hot encoding on Crypto Coin 
def crypto_one_hot_encoding(df):
  y_dummies = pd.get_dummies(df['Crypto'], prefix='Crypto', drop_first= False)
  df = pd.concat([df, y_dummies], axis=1)
  df.drop(['Crypto'], axis=1, inplace=True)
  # creating a additional column if the model is used for new coin.
  df['other_crypto'] =0
  return df

train_df = crypto_one_hot_encoding(train_df)
test_df = crypto_one_hot_encoding(test_df)

train_df['pct_change_1day'].describe()

test_df['pct_change_1day'].describe()

### Defining the Target Variable. 

We want to follow the classification approach and hence based on the "pct_change_2hour" we are creating 3 classes one class '0' when the returns are negative and '1' When the retruns are postive.

Finally, 2 when the returns beat the market value.

def create_target(df):
  market_RoR = 26.89
  market_RoR_1d = market_RoR/365
  df['Target'] = np.where(df['pct_change_1day']>0, 1,0)
  df['Target'] = np.where(df['pct_change_1day']>market_RoR_1d, 2,1)
  df['Target'][df['Target']==1] = np.where(df['pct_change_1day'][df['Target']==1]>=0, 1,0)
  return df

train_df = create_target(train_df)
test_df = create_target(test_df)

train_df['Target'].value_counts(normalize=True)

test_df['Target'].value_counts(normalize=True)

test_df.drop(['pct_change_1day'], axis=1, inplace=True) # droppping the column as we already extracted the target
train_df.drop(['pct_change_1day'], axis=1, inplace=True) # droppping the column as we already extracted the target

train_df.shape

test_df.shape

target = train_df['Target']
test_target = test_df['Target']

# Drop  columns

train_df.drop(['Target','Open Time','train_test',],axis=1,inplace=True)

test_df.drop(['Target','Open Time','train_test',],axis=1,inplace=True)

#plt.figure(figsize=(30,30))

#corr = train_df.drop(['Open','High', 'Low','Close','Crypto_ADA',
#       'Crypto_BTC', 'Crypto_ETC', 'Crypto_ETH', 'Crypto_LINK', 'Crypto_LTC',
#       'Crypto_TRX', 'Crypto_XLM', 'Crypto_XMR', 'Crypto_XRP', 'other_crypto'],axis=1).corr()
       
#sns.heatmap(corr,vmin=0, vmax=1, annot=True, cmap="YlGnBu")

# dropping the listof the columns 
# drop_columns = []
drop_columns = ['Open','Close']


if drop_columns:
  norm_train_df = train_df.drop(drop_columns,axis=1)
  norm_test_df = test_df.drop(drop_columns,axis=1)
else:
  norm_train_df = train_df
  norm_test_df = test_df

norm_train_df.columns

norm_train_df.shape

target.shape

# metrics for the classification
# from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
# def generate_model_report(y_actual, y_predicted):
#     print("Accuracy = " , accuracy_score(y_actual, y_predicted))
#     print("Precision = " ,precision_score(y_actual, y_predicted))
#     print("Recall = " ,recall_score(y_actual, y_predicted))
#     print("F1 Score = " ,f1_score(y_actual, y_predicted))

# Evaluation function

import sklearn
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
def generate_model_report(y_actual, y_predicted, metric_type):
    print("==================Printing the {} metrics===================".format(metric_type))
    if metric_type=='micro':
      print("Accuracy = " , round(accuracy_score(y_actual, y_predicted),4))
    print("Precision = " ,round(precision_score(y_actual, y_predicted, average=metric_type),4))
    print("Recall = " ,round(recall_score(y_actual, y_predicted, average=metric_type),4))
    print("F1 Score = " ,round(sklearn.metrics.f1_score(y_actual, y_predicted, average=metric_type),4))
    print("=====================================")
    return round(sklearn.metrics.f1_score(y_actual, y_predicted, average=metric_type),4)

!pip install tensorflow-addons

master_df = pd.DataFrame(columns=[daily['Crypto'].unique()])



master_df.loc['Dummy'] = 1
master_df.loc['XGboost'] = 1
master_df.loc['LGBM'] = 1
master_df.loc['LSTM'] = 1

master_df.loc['LSTM', 'BTC'] = 25

master_df



def model_predic(algo , norm_test_df):


  crypto_coins = daily['Crypto'].unique()


  for coin in crypto_coins:


    # print(coin)
    coin_index = norm_test_df[norm_test_df['Crypto_'+coin]==1]
    master_df.loc[algo, coin] = sklearn.metrics.f1_score(coin_index['pred'], coin_index['actual'], average='macro')
    
  # return master_df_x, master_df_y

## Dummy Predictor


 # Reading the file 
import pickle
temp_file_path = '/content/drive/MyDrive/MADS_23_DL_final_project/data/model_files/xgboost_hourly'
# loading the pickled model
with open(temp_file_path + '/dummy_f13_final.pkl', 'rb') as f:  # Python 3: open(..., 'wb')
    dummy_model = pickle.load(f)


norm_test_df['pred']=dummy_model.predict(norm_test_df.fillna(0))

norm_test_df['actual'] = test_target
model_predic('Dummy',norm_test_df)

master_df

## Xgboost

 # Reading the file 
import pickle
temp_file_path = '/content/drive/MyDrive/MADS_23_DL_final_project/data/model_files/xgboost_hourly'
# loading the pickled model
with open(temp_file_path + '/XGBoost_f13_final.pkl', 'rb') as f:  # Python 3: open(..., 'wb')
    XGboost_model = pickle.load(f)


norm_test_df['pred']=XGboost_model.predict(norm_test_df.fillna(0))

norm_test_df['actual'] = test_target
model_predic('XGboost',norm_test_df)

 # Reading the file 
import pickle
temp_file_path = '/content/drive/MyDrive/MADS_23_DL_final_project/data/model_files/xgboost_hourly'
# loading the pickled model
df2 = pd.read_csv(temp_file_path + "/dummy_cs.csv")

df2.head()

master_df.loc['XGboost'] = df2.loc[1].values

master_df

## LGBM

 # Reading the file 
import pickle
temp_file_path = '/content/drive/MyDrive/MADS_23_DL_final_project/data/model_files/LGBM_hourly'
# loading the pickled model
with open(temp_file_path + '/Tuned_lgbm_v13_final2.pkl', 'rb') as f:  # Python 3: open(..., 'wb')
    untuned_lgbm = pickle.load(f)



  

norm_test_df.columns


norm_test_df['pred']=untuned_lgbm.predict(norm_test_df[['High', 'Low', 'Volume', 'volatility_kcw', 'trend_cci', 'volume_adi',
       'momentum_ppo_hist', 'momentum_stoch', 'volatility_kcp', 'volume_em',
       'volatility_dcw', 'volume_vpt', 'volatility_bbw', 'Total_Value',
       'Total_Value_market', 'Value_Weight', '1_mon_lag', '2_mon_lag',
       '1_W_lag', '2_W_lag', '3_W_lag', '4_W_lag', '1_d_lag', '2_d_lag',
       '3_d_lag', '4_d_lag', '5_d_lag', '6_d_lag', '7_d_lag', '8_d_lag',
       'year', 'month_sin', 'month_cos', 'day_sin', 'day_cos', 'dayofweek_sin',
       'dayofweek_cos', 'hour_sin', 'hour_cos', 'Crypto_ADA', 'Crypto_BTC',
       'Crypto_ETC', 'Crypto_ETH', 'Crypto_LINK', 'Crypto_LTC', 'Crypto_TRX',
       'Crypto_XLM', 'Crypto_XMR', 'Crypto_XRP', 'other_crypto']].fillna(0))

norm_test_df['actual'] = test_target
model_predic('LGBM',norm_test_df)

temp_file_path = "/content/drive/MyDrive/MADS_23_DL_final_project/data/model_files/"
df2 = pd.read_csv(temp_file_path + "/dummy_.csv")

df2

master_df.loc['LSTM'] = df2.loc[3].values

master_df

master_df = master_df.T

import plotly.express as px 
fig = px.line(master_df, y=master_df.columns, x = master_df.index).update_layout(
    xaxis_title="Coins", yaxis_title="F1 Score(macro)",  width=600, height=400
)

# Show plot 

fig.show()

Observation: Overall the XGBoost and LGBM are doing better than rest of the algoritms. Since it is a classification problem the conventonal model are doing better than the Neural network Models.

Amoung the coins the model is able to best perform for the TRX coin.

## End of the Notebook
